Revert "Refactor hybrid search scripts into package"
This reverts commit 0a289311de57c51ac2309ca7b2cb0cc5c2803c09.
