+++ scripts/hybrid_embed.py
