+++ scripts/hybrid_expand.py
