+++ scripts/hybrid_filters.py
